src/requirements.txt
boto3==1.34.1
aiohttp==3.9.1
slack-sdk==3.26.1
python-dotenv==1.0.0
